Pass from double to float precision with relative corrections; Correction of implicit conversions occuring in multiple files

+++ Point2D.hpp
@@ -6,18 +6,18 @@
 namespace pf {
 struct Point2D
 {
-  double x;
-  double y;
+  float x;
+  float y;
 };
 
-inline double lunghezza(Point2D const& a)
+inline float lunghezza(Point2D const& a)
 {
   return std::sqrt(a.x * a.x + a.y * a.y);
 }
 
 inline Point2D normalizzazione(Point2D const& a)
 {
-  double len = lunghezza(a);
+  float len = lunghezza(a);
   if (len == 0) {
     return a;
   } else {
@@ -25,7 +25,7 @@ inline Point2D normalizzazione(Point2D const& a)
   }
 }
 
-inline double dot(Point2D const& a, Point2D const& b)
+inline float dot(Point2D const& a, Point2D const& b)
 {
   return a.x * b.x + a.y * b.y;
 }
@@ -40,19 +40,19 @@ inline Point2D operator-(Point2D const& a, Point2D const& b)
   return {a.x - b.x, a.y - b.y};
 }
 
-inline Point2D operator*(double u, Point2D const& a)
+inline Point2D operator*(float u, Point2D const& a)
 {
   return {u * a.x, u * a.y};
 }
 
-inline Point2D operator/(Point2D const& a, double u)
+inline Point2D operator/(Point2D const& a, float u)
 {
   return {a.x / u, a.y / u};
 }
 
-inline double distanza(Point2D const& a, Point2D const& b)
+inline float distanza(Point2D const& a, Point2D const& b)
 {
-  return std::sqrt(pow((a.x - b.x), 2) + pow((a.y - b.y), 2));
+  return std::sqrtf(powf((a.x - b.x), 2) + powf((a.y - b.y), 2));
 }
 
 } // namespace pf
