+++ statistics.hpp
@@ -5,17 +5,17 @@
 
 namespace pf {
 
-double mean_velocity(std::vector<Boid> const& stormo);
+float mean_velocity(std::vector<Boid> const& stormo);
 
-double mean_position(std::vector<Boid> const& stormo);
+float mean_position(std::vector<Boid> const& stormo);
 
-double standdev_velocity(std::vector<Boid> const& stormo);
+float standdev_velocity(std::vector<Boid> const& stormo);
 
-double standdev_position(std::vector<Boid> const& stormo);
+float standdev_position(std::vector<Boid> const& stormo);
 
-double mean_distance(std::vector<Boid> const& stormo);
+float mean_distance(std::vector<Boid> const& stormo);
 
-double standdev_distance(std::vector<Boid> const& stormo);
+float standdev_distance(std::vector<Boid> const& stormo);
 
 } // namespace pf
